:truck: :recycle: move common interfaces to base
