+++ libs/private/model/convs-mgr/micro-apps/base/README.md
@@ -1,0 +1,11 @@
+# private-model-convs-mgr-micro-apps-base
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test private-model-convs-mgr-micro-apps-base` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint private-model-convs-mgr-micro-apps-base` to execute the lint via [ESLint](https://eslint.org/).
