+++ libs/private/model/convs-mgr/micro-apps/base/src/index.ts
@@ -1,0 +1,5 @@
+export * from './lib/xAPI-statement.interface';
+export * from './lib/actor.interface';
+export * from './lib/course-participation.interface';
+export * from './lib/session.interface';
+export * from './lib/lms-launch-data.interface';
