+++ libs/private/model/convs-mgr/micro-apps/base/src/lib/actor.interface.ts
