+++ libs/private/model/convs-mgr/micro-apps/base/src/lib/course-participation.interface.ts
