+++ libs/private/model/convs-mgr/micro-apps/base/src/lib/lms-launch-data.interface.ts
@@ -1,0 +1,1 @@
+export interface LMSLaunchData { }
