+++ libs/private/model/convs-mgr/micro-apps/cmi5/src/index.ts
@@ -1,9 +1,3 @@
-export * from './lib/assignable-unit.interface';
-export * from './lib/context-template.interface';
-export * from './lib/launch-data.interface';
-export * from './lib/session.interface';
-export * from './lib/context-template.interface';
-export * from './lib/actor.interface';
-export * from './lib/xAPI-statement.interface';
-export * from './lib/au-status-types.enum';
-export * from './lib/course-participation.interface';
+export * from "./lib/context-template.interface";
+export * from "./lib/cmi5-launch-data.interface";
+
